chanstate: move revocation tlv adapters
Move the SparsePayHash TLV codec and HTLCEntry stream construction into
the revocation KV store file.

These helpers define the revocation log on-disk representation, so keep
them with the remaining revocation serialization while leaving the
domain file focused on types and constructors. This does not change the
encoding.

The stream constructor becomes an HTLCEntry method, which is a naming
change only.
