+++ chanstate/kv_revocation_log.go
@@ -8,11 +8,9 @@ import (
 	"math"
 
 	"github.com/lightningnetwork/lnd/kvdb"
+	"github.com/lightningnetwork/lnd/lntypes"
 	"github.com/lightningnetwork/lnd/tlv"
 )
-
-// This file contains the KV/TLV serialization helpers for revocation logs.
-// The domain types remain in revocation_log.go.
 
 var (
 	// revocationLogBucket is a sub-bucket under openChannelBucket. This
@@ -37,9 +35,9 @@ func RevocationLogBucketKey() []byte {
 }
 
 // PutRevocationLog uses the fields `CommitTx` and `Htlcs` from a
-// ChannelCommitment to construct a revocation log entry and saves them to disk.
-// It also saves our output index and their output index, which are useful when
-// creating breach retribution.
+// ChannelCommitment to construct a revocation log entry and saves them to
+// disk. It also saves our output index and their output index, which are
+// useful when creating breach retribution.
 func PutRevocationLog(bucket kvdb.RwBucket, commit *ChannelCommitment,
 	ourOutputIndex, theirOutputIndex uint32, noAmtData bool) error {
 
@@ -86,7 +84,8 @@ func PutRevocationLog(bucket kvdb.RwBucket, commit *ChannelCommitment,
 			continue
 		}
 
-		// Sanity check that the output indexes can be safely converted.
+		// Sanity check that the output indexes can be safely
+		// converted.
 		if htlc.OutputIndex > math.MaxUint16 {
 			return ErrOutputIndexTooBig
 		}
@@ -125,8 +124,67 @@ func FetchRevocationLog(log kvdb.RBucket,
 	return DeserializeRevocationLog(commitReader)
 }
 
-// htlcEntryToTlvStream converts an HTLCEntry record into a tlv representation.
-func htlcEntryToTlvStream(h *HTLCEntry) (*tlv.Stream, error) {
+// Record returns a tlv record for the SparsePayHash.
+func (s *SparsePayHash) Record() tlv.Record {
+	// We use a zero for the type here, as this'll be used along with the
+	// RecordT type.
+	return tlv.MakeDynamicRecord(
+		0, s, s.hashLen,
+		sparseHashEncoder, sparseHashDecoder,
+	)
+}
+
+// hashLen is used by MakeDynamicRecord to return the size of the RHash.
+//
+// NOTE: for zero hash, we return a length 0.
+func (s *SparsePayHash) hashLen() uint64 {
+	if bytes.Equal(s[:], lntypes.ZeroHash[:]) {
+		return 0
+	}
+
+	return 32
+}
+
+// sparseHashEncoder is the customized encoder which skips encoding the empty
+// hash.
+func sparseHashEncoder(w io.Writer, val interface{}, buf *[8]byte) error {
+	v, ok := val.(*SparsePayHash)
+	if !ok {
+		return tlv.NewTypeForEncodingErr(val, "SparsePayHash")
+	}
+
+	// If the value is an empty hash, we will skip encoding it.
+	if bytes.Equal(v[:], lntypes.ZeroHash[:]) {
+		return nil
+	}
+
+	vArray := (*[32]byte)(v)
+
+	return tlv.EBytes32(w, vArray, buf)
+}
+
+// sparseHashDecoder is the customized decoder which skips decoding the empty
+// hash.
+func sparseHashDecoder(r io.Reader, val interface{}, buf *[8]byte,
+	l uint64) error {
+
+	v, ok := val.(*SparsePayHash)
+	if !ok {
+		return tlv.NewTypeForEncodingErr(val, "SparsePayHash")
+	}
+
+	// If the length is zero, we will skip encoding the empty hash.
+	if l == 0 {
+		return nil
+	}
+
+	vArray := (*[32]byte)(v)
+
+	return tlv.DBytes32(r, vArray, buf, 32)
+}
+
+// toTlvStream converts an HTLCEntry record into a tlv representation.
+func (h *HTLCEntry) toTlvStream() (*tlv.Stream, error) {
 	records := []tlv.Record{
 		h.RHash.Record(),
 		h.RefundTimeout.Record(),
@@ -197,7 +255,7 @@ func SerializeRevocationLog(w io.Writer, rl *RevocationLog) error {
 func SerializeHTLCEntries(w io.Writer, htlcs []*HTLCEntry) error {
 	for _, htlc := range htlcs {
 		// Create the tlv stream.
-		tlvStream, err := htlcEntryToTlvStream(htlc)
+		tlvStream, err := htlc.toTlvStream()
 		if err != nil {
 			return err
 		}
