+++ chanstate/revocation_log.go
@@ -1,13 +1,10 @@
 package chanstate
 
 import (
-	"bytes"
-	"io"
 	"math"
 
 	"github.com/btcsuite/btcd/btcutil/v2"
 	"github.com/lightningnetwork/lnd/fn/v2"
-	"github.com/lightningnetwork/lnd/lntypes"
 	"github.com/lightningnetwork/lnd/lnwire"
 	"github.com/lightningnetwork/lnd/tlv"
 )
@@ -33,65 +30,6 @@ type SparsePayHash [32]byte
 // NewSparsePayHash creates a new SparsePayHash from a 32 byte array.
 func NewSparsePayHash(rHash [32]byte) SparsePayHash {
 	return SparsePayHash(rHash)
-}
-
-// Record returns a tlv record for the SparsePayHash.
-func (s *SparsePayHash) Record() tlv.Record {
-	// We use a zero for the type here, as this'll be used along with the
-	// RecordT type.
-	return tlv.MakeDynamicRecord(
-		0, s, s.hashLen,
-		sparseHashEncoder, sparseHashDecoder,
-	)
-}
-
-// hashLen is used by MakeDynamicRecord to return the size of the RHash.
-//
-// NOTE: for zero hash, we return a length 0.
-func (s *SparsePayHash) hashLen() uint64 {
-	if bytes.Equal(s[:], lntypes.ZeroHash[:]) {
-		return 0
-	}
-
-	return 32
-}
-
-// sparseHashEncoder is the customized encoder which skips encoding the empty
-// hash.
-func sparseHashEncoder(w io.Writer, val interface{}, buf *[8]byte) error {
-	v, ok := val.(*SparsePayHash)
-	if !ok {
-		return tlv.NewTypeForEncodingErr(val, "SparsePayHash")
-	}
-
-	// If the value is an empty hash, we will skip encoding it.
-	if bytes.Equal(v[:], lntypes.ZeroHash[:]) {
-		return nil
-	}
-
-	vArray := (*[32]byte)(v)
-
-	return tlv.EBytes32(w, vArray, buf)
-}
-
-// sparseHashDecoder is the customized decoder which skips decoding the empty
-// hash.
-func sparseHashDecoder(r io.Reader, val interface{}, buf *[8]byte,
-	l uint64) error {
-
-	v, ok := val.(*SparsePayHash)
-	if !ok {
-		return tlv.NewTypeForEncodingErr(val, "SparsePayHash")
-	}
-
-	// If the length is zero, we will skip encoding the empty hash.
-	if l == 0 {
-		return nil
-	}
-
-	vArray := (*[32]byte)(v)
-
-	return tlv.DBytes32(r, vArray, buf, 32)
 }
 
 // HTLCEntry specifies the minimal info needed to be stored on disk for ALL the
